red-brick: Work-around W5500 buffer config reset bug.
Sometimes the W5500 chip resets its buffer configuration to the defaults after some hours.
This breaks network connectivity completely, as no packets can be sent anymore.
The kernel log is then flooded with "kernel: socket_send(84000): free_buf_size 2048"
messages as well as "----PHY RESET----":

The driver writes one packet to the socket buffer, then issues a send command
and waits until the free buffer size is the expected 16384 bytes. But because of
the reset, that the chip does for some reason, the buffer has now only 2048 bytes
capacity, which results in an infinite loop in iinchip_socket_send.

Resetting the buffer configuration or other attempts to restore the ability
to send packets (e.g. the PHY RESET) did not help.

As the reported MTU is hardcoded to < 1500 anyway, we can live with the standard
buffer size of 2048 bytes.

To keep compatability with the W5200, we only skip the buffer configuration (and
fake the 2kb per socket assignment) if the chip is a W5500.